Various fixes

+++ addons/owens-godot-framework/tools/scripts/systems/pickup_system/throwable.gd
@@ -9,13 +9,20 @@ class_name OwenPickupThrowable
 # The force with which the object will be thrown.
 @export var throw_force : float = 1000.0
 
-# Called when the held and the player clicks.
+# Called when the object is held and the player clicks.
 func click_action():
+	# If the object is not held, return.
+	if parent_controller == null:
+		push_warning("OwenPickupThrowable: Object is not held, cannot throw.")
+		return
+	
 	# Get the mouse position and the player position.
 	var mouse_pos = get_global_mouse_position()
 	var player_pos = parent_controller.global_position
 	
 	# Drop the object.
+	# Note that drop will reset the parent controller to null.
+	# If you need to access variables from the parent controller, do it before calling drop.
 	parent_controller.drop()
 	super.click_action()
 	
@@ -23,19 +30,9 @@ func click_action():
 	var direction = (mouse_pos - player_pos).normalized()
 	apply_impulse(direction * throw_force)
 
-# Called when the object is picked up by the player.
-# Currently passes to the super class, but could be overridden to add additional behavior.
-func pick_up(_parent_controller : OwenPickupController):
-	super.pick_up(_parent_controller)
-
 # Called when the object is dropped by the player.
 func drop():
 	super.drop()
 	# Reset the linear and angular velocity of the object.
 	linear_velocity = Vector2.ZERO
 	angular_velocity = 0.0
-
-# Called when the object is next to be picked up by the player.
-# Currently passes to the super class, but could be overridden to add additional behavior.
-func next_to_pick_up(is_next : bool):
-	super.next_to_pick_up(is_next)
